project updated

+++ src/app/shared/issues.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { AngularFirestore, AngularFirestoreCollection } from '@angular/fire/firestore';
+import { AngularFirestore } from '@angular/fire/firestore';
 import { map } from 'rxjs/operators';
 import { Issue } from '../models/issue.model';
 import { firestore } from 'firebase/app';
@@ -13,25 +13,36 @@ export class IssuesService {
   //Create new issue
   //
   addIssue(data) {
-    return this.db
-      .collection('issues')
-      .add(
-        {
-          title: data.get('issueData.title').value,
-          tech: data.get('issueData.tech').value,          
-          dateStart: new Date(),
-          dateEnd: '-',
-          desc: {
-            summary: data.get('issueData.summary').value,
-            reproduce: data.get('issueData.reproduce').value,
-            expctRes: data.get('issueData.expctRes').value,
-            actlRes: data.get('issueData.actlRes').value
-          },
-          priority: data.get('issueData.priority').value,
-          status: 'Open',
-          notes: null
+    return new Promise((resolve, reject) => {
+      this.db.collection('issues')
+        .get()
+        .toPromise()
+        .then(() => {
+          const docRef = this.db.collection('issues')
+            .add(
+              {
+                title: data.get('issueData.title').value,
+                tech: data.get('issueData.tech').value,
+                dateStart: new Date(),
+                dateEnd: '-',
+                desc: {
+                  summary: data.get('issueData.summary').value,
+                  reproduce: data.get('issueData.reproduce').value,
+                  expctRes: data.get('issueData.expctRes').value,
+                  actlRes: data.get('issueData.actlRes').value
+                },
+                priority: data.get('issueData.priority').value,
+                status: 'Open',
+                notes: null
+              });
+          resolve("Success: Added Issue");
+        })
+        .catch(err => {
+          reject("Error: Adding issue failed");
         });
+    });
   }
+
 
   //Get all issues in realtime
   //
@@ -43,36 +54,40 @@ export class IssuesService {
   //
   getIssuesFiltered(priority: string, status: string) {
     return this.db.collection('issues', ref => ref.where('priority', '==', priority).where('status', '==', status)).snapshotChanges()
-    .pipe(
-      map(actions => actions.map(a => {
-        const data = a.payload.doc.data() as Issue;
-        const id = a.payload.doc.id;
-        return { id, ...data };
-      }))
-    );
+      .pipe(
+        map(actions => actions.map(a => {
+          const data = a.payload.doc.data() as Issue;
+          const id = a.payload.doc.id;
+          return { id, ...data };
+        }))
+      );
   }
-    
+
 
   //Get issues by the id paramater
   //Must get the firebase firestore document id using firestore.FieldPath.documentId()
+  //
   getIssuesById = (id) => {
     return new Promise((resolve, reject) => {
-      let data = this.db.collection('issues', ref => ref.where(firestore.FieldPath.documentId(), '==', id))
+      this.db.collection('issues', ref => ref.where(firestore.FieldPath.documentId(), '==', id))
         .get()
         .toPromise()
         .then((doc) => {
+          if (doc.empty) {
+            reject('No results found');
+          } else {
             let results = [];
-            doc.forEach( issue => {
+            doc.forEach(issue => {
               results.push({
                 id: issue.id,
                 data: issue.data()
-            })
-            resolve (results);
-          });
+              })
+              resolve(results);
+            });
+          }
         })
         .catch(err => {
-          console.log(err)
-          reject;
+          reject('No results found');
         });
     });
   }
@@ -81,26 +96,25 @@ export class IssuesService {
   //
   getIssuesByTech = (tech) => {
     return new Promise((resolve, reject) => {
-      let data = this.db.collection('issues', ref => ref.where('tech', '==', tech).where('status', '==', 'Open'))
+      this.db.collection('issues', ref => ref.where('tech', '==', tech).where('status', '==', 'Open'))
         .get()
         .toPromise()
         .then((doc) => {
           if (doc.empty) {
-            reject("ERROR: No documents were found");
+            reject('No results found');
           } else {
             let results = [];
-            doc.forEach( issue => {
+            doc.forEach(issue => {
               results.push({
                 id: issue.id,
                 data: issue.data()
               })
             })
-            resolve (results);
+            resolve(results);
           }
         })
         .catch(err => {
-          console.log(err)
-          reject;
+          reject('No results found');
         });
     });
   }
@@ -117,23 +131,53 @@ export class IssuesService {
       dateEnd = new Date();
     }
 
-    return this.db.collection('issues')
-      .doc(id)
-      .update(
-        {
-          tech: data.get('issueData.tech').value,
-          priority: data.get('issueData.priority').value,
-          status: data.get('issueData.status').value,
-          dateEnd: dateEnd,
-          notes: firestore.FieldValue.arrayUnion(data.get('issueData.notes').value)
+    return new Promise((resolve, reject) => {
+      this.db.collection('issues', ref => ref.where(firestore.FieldPath.documentId(), '==', id))
+        .get()
+        .toPromise()
+        .then(doc => {
+          if (doc.empty) {
+            reject("Error: Update Failed");
+          } else {
+            doc.docs.forEach((doc) => {
+              const docRef = this.db.collection('issues').doc(doc.id)
+                .update(
+                  {
+                    tech: data.get('issueData.tech').value,
+                    priority: data.get('issueData.priority').value,
+                    status: data.get('issueData.status').value,
+                    dateEnd: dateEnd,
+                    notes: firestore.FieldValue.arrayUnion(data.get('issueData.notes').value)
+                  });
+              resolve("Success: Updated Issue");
+            })
+          }
+        }).catch(err => {
+          reject("Error: Update Failed");
         });
+    });
   }
 
   //Delete issue using id param
   //
   deleteIssue(id: string) {
-    return this.db.collection("issues")
-      .doc(id)
-      .delete();
+    return new Promise((resolve, reject) => {
+      this.db.collection('issues', ref => ref.where(firestore.FieldPath.documentId(), '==', id))
+        .get()
+        .toPromise()
+        .then(doc => {
+          if (doc.empty) {
+            reject("Error: Delete Failed");
+          } else {
+            doc.docs.forEach((doc) => {
+              const docRef = this.db.collection('issues').doc(doc.id)
+                .delete();
+              resolve("Success: Deleted Issue");
+            })
+          }
+        }).catch(err => {
+          reject("Error: Delete Failed");
+        });
+    });
   }
 }
